[language][move] Replace the last uses of the token_match macro.
In the case of parse_pack_or_call(), it is clear what is expected next.
(It would be better if we could look up the symbol name and determine
whether it is a function or a struct, but that isn't in scope now.)
For parse_term(), I think it is better to just say that we expect an
"expression term" without listing all the possibilities.

Closes: #2454
Approved by: tzakian