[network] Remove nearly all proto from Discovery
This commit removes almost all proto from Discovery.

Note: This PR does some level of refactoring to move functions directly
into the structure definitions. It does not change the functionality of
the code that should not be in scope for this commit but we should
potentially discuss it and create an issue.

Closes: #2627
Approved by: phlip9