[GHA][LBT] handle expired GH token
We saw sporadic http error when LBT tries to post the cluster test
result on the PR. The error was about invalid GH token.  It turns out
the token expires after 60 minutes. If the workflow is queued up, the
token may have expired even before the job begins. In this case, LBT
needs to handle the http error.

Closes: #2684
Approved by: andll