[network] Introduce a new network interface wrapper
In order to move away from proto, this produces a new network interface
that wraps around lcs messages. It makes it slightly more convenient for
sending and receiving direct send and rpc.

This only shows how to do the modification on the test network, the next
commit will delete all the proto versions and then we can be rid of proto in
networking!

Closes: #2708
Approved by: zekun000