[storage] TreeCache supports bootstraping from PRE_GENESIS_VERSION
To support #2439 we need the state to be readable before geneiss version.
I'm defining PRE_GENESIS_VERSION = std::u64::max_value() and in progress of making our state tree capable of reaading from it.
This commit specifically tweeks the TreeCache so that it detects whether a pre genesis state is present when initialized with next_version == 0.

Closes: #3201
Approved by: lightmark