[language][vm] separate CompiledScript and CompiledModule in the file format
This is the beginning of the work to fully seperate scripts and modules.
While they share a few common tables for external dependencies, the
entry point (main) of a script is still very different from a full
fledged function definition in a module. This gets it simplified by
replacing the full function definition with only the fields required,
also eliminating the need for a few other table entries.

Closes: #3500
Approved by: dariorussi