[crypto] X25519 PrivateKey now wraps x25519-dalek StaticSecret
There is no good reason for us to not have this be an actual dalek type
(in place of a [u8; 32]), since we never use the byte representation.

Furthermore, this implements a number of handy functions
(for example diffie_hellman), that allow us to avoid movinging back and
forth between our wrappers and the inner types.

Closes: #3576
Approved by: davidiw