[move-lang] Use the term "identifier" instead of "name" in the parser
The term "identifier" is well known and used by other compilers like
Clang and Rust, both internally and in diagnostic messages. We should
do the same for Move. (I would also like to repurpose the "name" term
for what we now call a "ModuleAccess", but let's take this one step
at a time....)

Closes: #3651
Approved by: sblackshear