[move-lang] Specify input files as arguments, not options
This aligns the Move compiler with the prover's command-line syntax,
where the input files are specified as arguments instead of options.
This is also how most other compilers specify input files.

Rename the "--dependencies" option to "--dependency". This option
can be used multiple times to produce a vector of dependencies but
each option specifies a single dependency (or possibly a single
directory of dependency files, but still singular).

Closes: #3970
Approved by: wrwg