[crypto] Expose the serialization seed through CryptoHasher
This allow access to the initial sha3-hashed bytes of the LIBRA_HASHING_PREFIX + LCS trace_name for a struct that implements CryptoHasher.

It also exposes it for the legacy macros.

Since the CryptoHash implementations are not explicit in the code base these days, there's a low chance of actual misuse of this machinery.

Closes: #4073
Approved by: matbd