[move-prover] Removing redundant type assumptions ==> new havocs
We had a while ago added a redundant type assumption. Basically whenever we had an assignment `x := y` the type of `x` was assumed again. This was done to hunt a bug but forgotten to remove.

It appeared that this masked an additional problem with havoc in loops. After removing this, multiple loop invariants aren't working again. It is an instance of the same problem we have seen before, namely that not all assumptions about value should be havoced. Here it appears to be the types.

This PR removes the redundant type assumptions, and deactivates the failing tests. This applies to functional/{loops.move, specs_in_fun_ref.move). Those tests should be reactivated in a subsequent PR when this is fixed.

Notice that this was blocking PR #4304. This PR run into unexpected havoc errors in those tests after removing redundant temporaries, and therefore not longer benefitting from the masking of the havoc. Also notice
that likely more complicated code then the current tests would have also hit this problem, even if we would leave the redundant assumptions.

Closes: #4456
Approved by: emmazzz