[scripts] Added convenience script start_cli_swarm.sh
Recently swarm was changed so, that --libra-node and --cli-path are
madatory, raising two issues:
    - there were no easy way to lauch swarm with CLI, because the
      documentation was out-of-date, and
    - the new way of lauching would not fit well to the documentation
      because the user might have "debug" or "release" version.

This script solves both problems by making lauching the swarm nice and
easy. It also forwards aguments after "--" straight to the swarm, so
user can configure their swarm if so desired.

Default is to build and use "debug" build, but "release" build can be
specified by issuing "-r" or "--release".