[network] Simplify Network Healthcheck
The network healtcheck was randomly picking a peer to healthcheck.  If
there are a large number of peers, the likelihood of healthchecking a
peer at a regular rate diminishes.  It was as strange behavior that
really doesn't healthcheck at a regular rate.  This now, healthchecks
all peers at a regular rate, creating different nonces for each.  The
nonces are to prevent spoofs of health check responses.

Closes: #6712