[prover] separate functions for is_native and is_intrinsic
Previously, is_native conflated these two concepts, which frustrates analysis clients that want to treat these differently.
The motivation for this change is to support the read/write set analysis, which needs to analyze the body of functions marked as intrinsic.

The PR changes all uses of `is_native` in bytecode_translator.rs and spec_instrumentation.rs to `is_native && is_intrinsic`. There are more uses of `is_native` in some of the other bytecode analyses/transformations (e.g., borrow_analysis), but I left them as is because it seems like they also want to ignore the intrinsic pragma

Closes: #7272