[move-prover] use invariant instead of assert to mark loop invariant
We need a special mark for loop invariant propositions because those
will be translated to both asserts and assumes (under the loop-to-dag
pass). To prepare for this transition, this commit additionally allows
`invariant <exp>` to be specified in a spec block in function code but
transforms the `invariant <exp>` into an `assert <exp>` in the loop
anaylsis pass.

Note that this commit is compatible with the existing practice: one may
continue to use `assert` to specify loop invariants in the header block
and it will still work.