[move-prover] Representing structs as ADTs instead of vectors.
This PR again forks the Boogie backend, this time into a new `boogie-backend-exp`. The new backend implements Move structs as Boogie datatypes instead of ADTs.

The advantage of this representation is that no boxing of the field data is needed, as is the case for the vector representation `Vec $Value`. (Boxing is still required if the field has a generic type.) Moreover, as long as structs do not contain vectors, extensional equality holds. However, the `$Value` type is still needed with this PR, since we do not (yet) fully monomorphize. This leads to a representation of `$Value` as such:

```
type {:datatype} $Value;
// Predefined constructors
fun {:constuctor} $Integer(i: int): $Value
..

// Generated constructors
fun {:constructor} $Struct1_box(s: $Struct1): $Value;
fun {:constructor} $Struct2_box(s: $Struct2): $Value;
.. // and so on for all structs part of the verification problem
```

Benchmak data in the new lab `lab/struct-as-adt` shows that the ADT representation is slightly slower *but* has less timeouts. This might be because update of fields in ADT representation might be more expensive than update on vectors, as long as they are represented as Boogie/smt arrays. However, a final verdict cannot be made until the monomorphization work is finished and all usage of the `$Value` type has been eliminated.

For now, the new struct representation is not on by default but guarded by the flag `--bexp`. All tests are executed for both the default and the `bexp` feature (and pass).

Note that this is a larger PR because many rewrites of prelude and code generation had to be done. We can unfortunately not see a diff from forking of the boogie backend. The PR was shelved a week ago or so to first get the new prover-lab in so systematic benchmarking could be done.

Closes: #8177