[move-prover] Revert "Revert "Representing structs as ADTs instead of vectors.""
This reverts commit 72a378e76ea028ca59d2ab37b7ca885ef94b57f9 after fixing the underlying issue which caused tests to fail.

The issue was that after moving dep analysis int move-lang, lots of modules are included even though they are never used. This led to a large case distinction for $IsEqual_struct in the new ADT boogie backend, which in turn generated prover timeouts. We are now optimizing the implementation of this function by analyzing which structs are actually needed for a given VC condition.

Closes: #8208