[move-lang] Add unit test support to the Move compiler
This PR adds support for test attributes (#[test(..)], #[test_only], and #[expected_failure)
to the Move compiler.

In non-test mode all module members that are annotated as #[test] or #[test_only] will be filtered out from compilation before expansion.

In test mode, the module is unchanged. At the end of compilation, we
then create a test plan for each module in the program compiled. This is
then used by the unit testing driver that is in the commit above this.

Need to add in additional tests for other types of module members.