[read/write set analysis] fix crash when updating return value via join
The analysis had an assertion that was meant to guard against updating `Return` variables more than once (since that should never happen). However, the assertion was overzealous in that although a `Return` can only be assigned once, the value associated with a `Return` can be updated via a join (e.g., when a procedure returns from several different control locations).

This PR adds a test that crashed the old analysis + removes the overzealous assertion.

Closes: #8258