[move stdlib] Capability-based secure storage.
This PR introduces a new module (in the nursery), called *Vault*, which implements a secure memory of some content which can only be operated
on if authorized by a signer. Authorization is managed by [*capabilities*](https://en.wikipedia.org/wiki/Capability-based_security) in the proper sense as used in
security engineering. Capabilities are unforgeable tokens which represent the successful authorization of access based on a signer, and which can be passed
on to other functions which can perform the operations enabled by the capability, without the need to have access to the signer themselves.

The vault comes with 4 kinds of capabilities:

- Read: allows to read the content.
- Modify: allows to read and write the content.
- Delegate: allows to delegate access (of one or more capabilities) to another signer. Delegation can be revoked.
- Transfer: allows to transfer ownership of the vault to a new signer.

Reading and writing of the vault is enabled by using a wrapper object (called an accessor) which moves the value temporarily out of the vault and allows to created references to it. The accessor
needs to be explicitly released. This is the current workaround for that we can't return references to global memory.

This PR also enables documentation generation for modules in the nursery. The generated doc for the new module is [here]().

There are currently no specs or tests, which need to be added in later PRs.

Closes: #8396